+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Change Log
 
+### v1.0.1(Mar 30, 2018)
+* add SendBirdDesk.init().
+* add SendBirdDesk.isDeskChannel(channel).
+* add TypeScript interface - SendBird.Desk.d.ts.
+* add console logger in debug mode.
+
 ### v1.0.0-zendesk(Mar 16, 2018)
 * Zendesk-integrated version release.